requirements.txt
asgiref==3.8.1
Django==5.0.3
lxml==5.1.1
pillow==10.2.0
python-pptx==0.6.23
sqlparse==0.4.4
XlsxWriter==3.2.0
